+++ docs/whatsnew.md
@@ -5,6 +5,7 @@
 - fix : values of arrays with just one element weren't return as list
 - fix : correct .dataType of multi-dimensional structure arrays
 - fix : correct repr of Variable (VT was incorrect)
+- fix : bug reading SNMP variables
 - new property 'isStructure' for Variable object
 
 ## V1.2.9